+++ core-ng-search/src/main/java/core/framework/search/SearchRequest.java
@@ -3,11 +3,13 @@ package core.framework.search;
 import co.elastic.clients.elasticsearch._types.SearchType;
 import co.elastic.clients.elasticsearch._types.SortOptions;
 import co.elastic.clients.elasticsearch._types.aggregations.Aggregation;
+import co.elastic.clients.elasticsearch._types.mapping.RuntimeField;
 import co.elastic.clients.elasticsearch._types.query_dsl.Query;
 import core.framework.util.Lists;
 import core.framework.util.Maps;
 
 import javax.annotation.Nullable;
+import java.io.StringReader;
 import java.util.List;
 import java.util.Map;
 
@@ -16,6 +18,7 @@ import java.util.Map;
  */
 public class SearchRequest {
     public final Map<String, Aggregation> aggregations = Maps.newHashMap();
+    public final Map<String, RuntimeField> runtimeFields = Maps.newHashMap();
     public final List<SortOptions> sorts = Lists.newArrayList();
     @Nullable
     public String index;
@@ -31,5 +34,13 @@ public class SearchRequest {
 
     public void trackTotalHits() {
         trackTotalHitsUpTo = Integer.MAX_VALUE;
+    }
+
+    public void withJSON(String source) {
+        var request = co.elastic.clients.elasticsearch.core.SearchRequest.of(b -> b.withJson(new StringReader(source)));
+        if (request.query() != null) query = request.query();
+        if (request.aggregations() != null) aggregations.putAll(request.aggregations());
+        if (request.runtimeMappings() != null) runtimeFields.putAll(request.runtimeMappings());
+        if (request.sort() != null) sorts.addAll(request.sort());
     }
 }